Added reusable function for updating dynamo calls
